refactor: Split out PlayerSkillsEvent
